Rename picker to choose
