+++ main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"os"
 
+	"github.com/anttiharju/vmatch/pkg/choose"
 	"github.com/anttiharju/vmatch/pkg/interrupt"
-	"github.com/anttiharju/vmatch/pkg/picker"
 )
 
 func main() {
 	go interrupt.Listen(os.Interrupt)
 
-	exitCode := picker.SelectWrapper(os.Args[1:])
+	exitCode := choose.Wrapper(os.Args[1:])
 	os.Exit(exitCode)
 }
