+++ pkg/choose/choose.go
@@ -1,4 +1,4 @@
-package picker
+package choose
 
 import (
 	"github.com/anttiharju/vmatch/pkg/wrapper/language"
@@ -9,7 +9,7 @@ func firstArgIs(arg string, args []string) bool {
 	return len(args) > 0 && args[0] == arg
 }
 
-func SelectWrapper(args []string) int {
+func Wrapper(args []string) int {
 	if firstArgIs("go", args) {
 		wrappedLanguage := language.Wrap("go")
 		exitCode := wrappedLanguage.Run(args[1:])
